Move font scale hook to hooks dir
