+++ components/eventForm/DismissButton.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { Alert, TouchableOpacity } from "react-native"
 import { useEventFormContext } from "./EventForm"
 import { MaterialIcon } from "@components/common/Icons"
-import { FontScaleFactors } from "@lib/FontScale"
+import { FontScaleFactors } from "@hooks/Fonts"
 
 /**
  * A dismiss button for `EventForm`.
